requirements.txt
pyacoustid
flask
flask-sockets
gevent
gevent-websocket
gunicorn
pbkdf2
psycopg2-binary
requests
rply
alembic
sqlalchemy
sqlalchemy-utils
tinytag
python-magic
pillow
ftfy
audioread
